fix: move styles from page
