+++ src/components/faq-accordion/styles.ts
