+++ src/pages/faq/index.tsx
@@ -1,11 +1,11 @@
 import Head from 'next/head';
+
 import { Background, HeadContainer, Title } from 'src/components/appointments/styles';
 import FaqAccordion from 'src/components/faq-accordion/FaqAccordion';
+import { StyledContainer } from 'src/components/faq-accordion/styles';
 import { PrivateRoute } from 'src/components/private-route/PrivateRoute';
-
 import MainHeader from 'src/components/reusable/header/MainHeader';
 import { useLocales } from 'src/locales';
-import { StyledContainer } from './styles';
 
 export default function FaqPage(): JSX.Element | null {
   const { translate } = useLocales();
